+++ tools.h
@@ -7,6 +7,6 @@
 
 time_t get_file_time(const char* file_name);
 int64_t get_time_ms();
-void sleep_ms(int64_t ms);
+void sleep_ms(int ms);
